Checked symbolic links have been added

+++ src/chfiles2deb.sh
@@ -2,21 +2,25 @@
 
 # Функция вывода справки
 usage() {
-    echo "Использование: $0 [путь_для_поиска] <имя_пакета.deb> <флаги_времени_find>"
+    echo "НАЗНАЧЕНИЕ:"
+    echo "  Скрипт находит новые и измененные файлы с помощью утилиты find"
+    echo "  и собирает из них готовый DEB-пакет с сохранением структуры каталогов."
+    echo ""
+    echo "Использование: $0 [путь_для_поиска] <имя_пакета.deb> <флаги_времени_find> [-k | --keep]"
     echo "   или: $0 -h | --help"
-    echo "Находит новые и измененные файлы и собирает из них готовый DEB-пакет."
     echo ""
     echo "Обязательные параметры:"
     echo "  флаги_времени_find  Любые временные флаги утилиты find (например: -mmin -5, -mtime -1)"
     echo "  <имя_пакета.deb>    Имя DEB-пакета"
     echo ""
     echo "Необязательные параметры:"
-    echo "  [путь_для_поиска]     Каталог для сканирования (по умолчанию текущий каталог) "
+    echo "  [путь_для_поиска]     Каталог для сканирования (по умолчанию текущий каталог '.')"
+    echo "  -k, --keep          Не удалять временный каталог с файлами после сборки"
     echo ""
     echo "Примеры вызова:"
     echo "  $0 -mmin -10 foo.deb"
-    echo "  $0 /usr/local foo.deb -mtime -1"
-    echo "  $0 /usr/local foo.deb -newermt 09:15:00"
+    echo "  $0 /var/www /tmp/web_site.deb -mtime -1 --keep"
+    echo "  $0 /usr/local foo.deb -newermt \"2026-06-19 09:15:00\""
     exit 1
 }
 
@@ -24,6 +28,7 @@ SRC_DIR=""
 DEB_FILE=""
 TIME_PARAMS=()
 POSITIONAL_ARGS=()
+KEEP_BUILD_DIR=false 
 
 # Цикл разбора аргументов командной строки
 while [[ $# -gt 0 ]]; do
@@ -31,8 +36,11 @@ while [[ $# -gt 0 ]]; do
         -h|--help)
             usage
             ;;
+        -k|--keep)
+            KEEP_BUILD_DIR=true
+            shift 1
+            ;;
         -mmin|-mtime|-amin|-atime|-cmin|-ctime|-newer*)
-            # Аргумент временных ограничений
             TIME_PARAMS+=("$1" "$2") 
             shift 2
             ;;
@@ -41,7 +49,6 @@ while [[ $# -gt 0 ]]; do
             usage
             ;;
         *)
-             # Аргумент(ы) пути поиска и/или архива с измененными файлами
             POSITIONAL_ARGS+=("$1")
             shift 1
             ;;
@@ -53,6 +60,7 @@ if [ ${#TIME_PARAMS[@]} -eq 0 ] || [ ${#POSITIONAL_ARGS[@]} -eq 0 ]; then
     usage
 fi
 
+# Логика распределения путей (архив и каталог)
 if [ ${#POSITIONAL_ARGS[@]} -eq 1 ]; then
     DEB_FILE="${POSITIONAL_ARGS[0]}"
     SRC_DIR="."
@@ -65,7 +73,6 @@ SRC_DIR=$(realpath "$SRC_DIR")
 DEB_FILE=$(realpath "$DEB_FILE")
 SCRIPT_PATH=$(realpath "$0")
 
-# Получение имени пакета без пути и расширения .deb
 PKG_NAME=$(basename "$DEB_FILE" .deb)
 
 # Создание уникальной временной директории со случайным именем
@@ -74,10 +81,16 @@ BUILD_DIR=$(mktemp -d -t deb_build_XXXXXX)
 # Функция автоматической очистки при выходе или сбое
 cleanup() {
     if [ -d "$BUILD_DIR" ]; then
-        rm -rf "$BUILD_DIR"
+        if [ "$KEEP_BUILD_DIR" = false ]; then
+            rm -rf "$BUILD_DIR"
+        else
+            if [ -d "$BUILD_DIR/DEBIAN" ]; then
+                echo "Временный каталог сохранен: $BUILD_DIR"
+                echo "Для сборки пакета выполните команду: dpkg-deb --build \"$BUILD_DIR\" \"$DEB_FILE\""
+            fi
+        fi
     fi
 }
-# Перехватываем обычный выход (EXIT), Ctrl+C (SIGINT) и завершение процесса (SIGTERM)
 trap cleanup EXIT SIGINT SIGTERM
 
 EXCLUDES=(
@@ -91,9 +104,8 @@ EXCLUDES=(
     -o -path "$BUILD_DIR"
 )
 
-# Поиск и копирование измененных файлов во временный каталог с сохранением структуры путей
+# Поиск и копирование измененных файлов и символических ссылок
 find "$SRC_DIR" \( "${EXCLUDES[@]}" \) -prune -o \( -type f -o -type l \) "${TIME_PARAMS[@]}" -print0 | tar --null -T - -cf - | tar -xf - -C "$BUILD_DIR"
-
 
 # Проверка наличия файлов
 if [ -z "$(ls -A "$BUILD_DIR")" ]; then
@@ -101,19 +113,36 @@ if [ -z "$(ls -A "$BUILD_DIR")" ]; then
     exit 0
 fi
 
+# ПРОВЕРКА СИМВОЛИЧЕСКИХ ССЫЛОК
+echo "Проверка целостности символических ссылок внутри сборки..."
+while read -r -d '' link_path; do
+    target=$(readlink "$link_path")
+    
+    if [[ "$target" = /* ]]; then
+        full_target_path="$BUILD_DIR$target"
+    else
+        link_dir=$(dirname "$link_path")
+        full_target_path=$(realpath -m "$link_dir/$target")
+    fi
+
+    if [ ! -e "$full_target_path" ]; then
+        display_link="${link_path#$BUILD_DIR}"
+        echo "Предупреждение: Ссылка [$display_link] ведет на отсутствующий в пакете объект [$target]"
+        KEEP_BUILD_DIR=true
+    fi
+done < <(find "$BUILD_DIR" -type l -print0)
+
 # Создание обязательных метаданных пакета (DEBIAN/control)
 mkdir -p "$BUILD_DIR/DEBIAN"
 cat << EOF > "$BUILD_DIR/DEBIAN/control"
 Package: $PKG_NAME
 Version: 1.0.$(date +%Y%m%d%H%M)
 Architecture: all
-Maintainer: BackupScript <admin@localhost>
+Maintainer: Sergey Maksimov <m6v@mail.ru>
 Description: Автоматический пакет обновлений измененных файлов
 EOF
 
 # Сборка DEB-пакета
-# Если нужно, чтобы внутри .deb пакета файлы принадлежали конкретному пользователю (например, root:root),
-# в блок сборки перед dpkg-deb добавить флаг --root-owner-group
 echo "Сборка пакета $DEB_FILE..."
 dpkg-deb --build "$BUILD_DIR" "$DEB_FILE"
